added left rever responces.

+++ NFCq/src/com/chernicka/model/User.java
@@ -17,8 +17,8 @@ public class User {
     private String key;
     private String score;
 
-    private int latitude;
-    private int longtitude;
+    private double latitude;
+    private double longtitude;
 
     public User(){}
 
@@ -42,11 +42,11 @@ public class User {
         return score;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public int getLongtitude() {
+    public double getLongtitude() {
         return longtitude;
     }
 
@@ -70,11 +70,11 @@ public class User {
         this.score = score;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongtitude(int longtitude) {
+    public void setLongtitude(double longtitude) {
         this.longtitude = longtitude;
     }
